+++ Cargo.toml
@@ -14,6 +14,7 @@ edition = "2021"
 [dependencies]
 db_error = { path = "crates/db_error", version = "<=0.1.0" }
 serialize = { path = "crates/serialize", version = "<=0.1.0" }
+storage = { path = "crates/storage", version = "<=0.1.0" }
 
 [dev-dependencies]
 test_file = { path = "crates/test_file", version = "<=0.1.0" }
