+++ crates/storage/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "storage"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+db_error = { path = "../db_error", version = "<=0.1.0" }
+serialize = { path = "../serialize", version = "<=0.1.0" }
+
+[dev-dependencies]
+test_file = { path = "../test_file", version = "<=0.1.0" }
