+++ crates/storage/src/lib.rs
@@ -1,0 +1,15 @@
+mod file_storage;
+mod file_storage_data;
+mod storage;
+mod storage_data;
+mod storage_drop;
+mod storage_record;
+mod storage_record_with_index;
+mod storage_records;
+mod write_ahead_log;
+mod write_ahead_log_record;
+
+pub use file_storage::FileStorage;
+pub use file_storage_data::FileStorageData;
+pub use storage::Storage;
+pub use storage_data::StorageData;
