+++ src/graph/storage_dictionary.rs
@@ -2,14 +2,14 @@ use super::dictionary_data_storage::DictionaryDataStorage;
 use super::dictionary_data_storage_indexes::DictionaryDataStorageIndexes;
 use super::dictionary_impl::DictionaryImpl;
 use super::dictionary_value::DictionaryValue;
-use crate::storage::FileStorageData;
 use crate::storage::StableHash;
-use crate::storage::Storage;
-use crate::storage::StorageData;
 use crate::storage::StorageHashMultiMap;
 use crate::storage::StorageVec;
 use crate::DbError;
 use serialize::Serialize;
+use storage::FileStorageData;
+use storage::Storage;
+use storage::StorageData;
 
 pub(crate) type StorageDictionary<T, Data = FileStorageData> =
     DictionaryImpl<T, DictionaryDataStorage<T, Data>>;
@@ -95,7 +95,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::storage::FileStorage;
+    use storage::FileStorage;
     use test_file::TestFile;
 
     #[derive(Clone, Default, Eq, PartialEq)]
