+++ src/storage/storage_hash_map.rs
@@ -1,13 +1,13 @@
-use super::file_storage_data::FileStorageData;
 use super::hash_map_data_storage::HashMapDataStorage;
 use super::hash_map_impl::HashMapImpl;
 use super::hash_map_key_value::HashMapKeyValue;
 use super::stable_hash::StableHash;
-use super::storage_data::StorageData;
-use super::Storage;
 use crate::DbError;
 use serialize::Serialize;
 use std::hash::Hash;
+use storage::FileStorageData;
+use storage::Storage;
+use storage::StorageData;
 
 pub(crate) type StorageHashMap<K, T, Data = FileStorageData> =
     HashMapImpl<K, T, HashMapDataStorage<K, T, Data>>;
@@ -93,7 +93,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::storage::file_storage::FileStorage;
+    use storage::FileStorage;
     use test_file::TestFile;
 
     #[test]
